// services/games/src/domain/bet/bet-status.enum.ts
export enum BetStatus {
  PENDING_DEBIT = 'PENDING_DEBIT',
  ACCEPTED = 'ACCEPTED',
  REJECTED = 'REJECTED',
  CASHED_OUT = 'CASHED_OUT',
  LOST = 'LOST',
}

export enum BetRejectedReason {
  INSUFFICIENT_FUNDS = 'INSUFFICIENT_FUNDS',
  DUPLICATED_BET = 'DUPLICATED_BET',
  ROUND_NOT_BETTING = 'ROUND_NOT_BETTING',
  UNKNOWN = 'UNKNOWN',
}